test(research): extract PDF grouping layout fixtures
